Added Left-Right Movement Functionality to Paddle
The 'A' and 'D' keys can now be used to move the paddle left or
right by ten pixels per second. However, there's no edge detection
so the paddle just continues off screen in either direction.
Also, WSAD ball control was removed to make this work. If WSAD
ball control is really THAT necessary of a feature it can be undone
and controlled in some other way.