Changed Brick Color and Removed Bounce Edge
The bricks are now red to make it easier to see when the console
is open. The bottom bounce edge of the ball has been removed.
Now, when the ball flies below the paddle it doesn't bounce off
the bottom of the screen but instead keeps going downwards.
This is when a death would occur which will soon be implemented.