Added Game Over Mechanic
When the ball hits the bottom of the screen a game over sequence
occurs where the player can't control anything for 3 seconds.
After that interval, the player may hit any key to start again
which resets the ball and bounce counter. The bricks don't reset
right away though because of a planned lives feature.