Added Lives Mechanic and Removed Excessive Explosions
When the ball reaches the bottom of the screen, it now causes the
player to loose a life. When the player reaches 0 lives, the game
ends, giving the player an option to restart. Soon, the bricks
will reset. Also, explosions that aren't part of brick destruction
were removed because too many explosions were happening!