Added Brick Health Scaling Based on Level
The bricks now increase in health based on location and level.
In level 1, the bricks increase by 1 health every 30 bricks,
resulting in the first 30 bricks having 1 health and the last 30
having 2. In level 2, the health increases by 1 every 20 bricks,
meaning the first 20 have 1 health, the second 20 have 2 health,
and the last 20 have 3 health. Lastly, in level 3, the brick
health increases every row, so the first row has 1 health,
the second row has 2, all the way up to the last 10 bricks with
6 health.