Added Color Changing Bricks
The bricks now take on a color based on their health ranging from
one to six. Each health level is tied to a brick color.
The health corellates to colors like so:
1 -> dark blue
2 -> turqoise
3 -> green
4 -> yellow
5 -> orange
6 -> red

When a brick is hit, it's color is updated based on its new
health if it still hasn't been destroyed. This makes it
much easier to tell how much health a given brick has.