Added Level Transitions
When all bricks are destroyed in a given level, a level victory
message is displayed to the player. Soon, the player is prompted
for a key press like they'd be if they lost a life. When the player
presses a key, the bricks are reset and scaled up in health and a
new level begins.