Added God Mode Command to HUD
The command "godmode" enables/disables godmode, a mode where the
player is invincible. Basically, the ball just bounces off the bottom
of the screen instead of causing a life loss / game over.
Typing the command into the HUD enables or disables it depending
on the previous setting, god mode is off by default. It will
say the result in the HUD.