Added instant death command to HUD
When the player types in "lose" to the HUD, the player loses a life
or even the game if they only have 1 life left.
This command basically triggers the losing state instantly, the
opposite of the winlevel command. This allows the player to quickly
reach the game over state if they so desire.