Added Basic Splash Screen
The screen currently displays unrelated stand-in text when the game
is booted up. This text acts as the splash screen and goes away
after five or so seconds. This screen is also skippable using the
ESC key. After pressing ESC or waiting the time, the screen switches
to the load screen which is skippable using SPACE or the player can
just wait for it to end. Then the game starts from there.
Soon, the stand-in text will be replaced with actually relevant text
for the splash screen.