Fixed Brick Health Bug
Now, the bricks properly act like the health levels they should be
at. For instance, when the turqoise brick is hit it becomes a dark
blue brick instead of ceasing to exist. This was fixed by simply
checking if the health was bigger than 0 instead of 1. So, an off
by one error. Fun.