Added tiny gravitational effect on ball
A tiny 0.05 unit gravity affect is now constantly applied to the ball.
This makes it so if the ball gets stuck in the edge case of moving
exculusively side to side, it is now not stuck like that, being
capable of moving down slowly due to gravity affecting it.
It's also a realistic compensation because gravity is a real physical
thing.