Replaced lose command with die command
The die command acts similarly to the lose command of old but it's
better. Basically, typing die into the HUD terminal causes the
player to lose a life or enter game over if they have no lives
left after the command. It works better than the previous lose
command and preserves godmode.