Added setlevel command to HUD terminal
The command setlevel can be used to set the level of the game
from range 1 to 3. If the given input level input is out of range
1 to 3 inclusive, then invalid level is returned and nothing
happens. This command is useful for seeing the three levels of
the game without winning a bunch. It's also useful if the player
needs to reset a later level without having to die and win back
to that level.